refactor: 이전단계 문서 내용 제거, request 패키지 이동
