+++ README.md
@@ -61,32 +61,4 @@ This project is [MIT](https://github.com/next-step/atdd-subway-service/blob/mast
 - [X] 인증 - 내 정보 조회 기능 완성하기
 - [X] 인증 - 즐겨 찾기 기능 완성하기
 
-#### 미션 수행 순서
-
-인수 테스트 성공 시키기
-
-- mock 서버와 dto를 정의하여 인수 테스트 성공 시키기
-
-- 기능 구현
-
-```
-TDD의 방향보다 테스트를 통해 구현할 기능을 명세하는것과 리팩터링이 더 중요합니다!
-두 방향성을 모두 사용해보시고 테스트가 협력 객체의 세부 구현에 의존하는 경우(가짜 협력 객체 사용)와 테스트 대상이 협력 객체와 독립적이지 못하고 변경에 영향을 받는 경우(실제 협력 객체 사용)를 모두 경험해보세요 :)
-```
-
-Outside In 경우
-
-- 컨트롤러 레이어 구현 이후 서비스 레이어 구현 시 서비스 테스트 우선 작성 후 기능 구현
-- 서비스 테스트 내부에서 도메인들간의 로직의 흐름을 검증, 이 때 사용되는 도메인은 mock 객체를 활용
-- 외부 라이브러리를 활용한 로직을 검증할 때는 가급적 실제 객체를 활용
-- Happy 케이스에 대한 부분만 구현( Side 케이스에 대한 구현은 다음 단계에서 진행)
-
-Inside Out 경우
-
-- 도메인 설계 후 도메인 테스트를 시작으로 기능 구현 시작
-- 해당 도메인의 단위 테스트를 통해 도메인의 역할과 경계를 설계
-- 도메인의 구현이 끝나면 해당 도메인과 관계를 맺는 객체에 대해 기능 구현 시작 ex) 경로 조회를 수행하는 도메인 구현 예시
-    1. PathFinder 라는 클래스 작성 후 경로 조회를 위한 테스트를 작성
-    2. 경로 조회 메서드에서 Line을 인자로 받고 그 결과로 원하는 응답을 리턴하도록 테스트 완성
-    3. 테스트를 성공시키기 위해 JGraph의 실제 객체를 활용(테스트에서는 알 필요가 없음)
