+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,8 +1,8 @@
 package nextstep.subway.favorite.application;
 
-import nextstep.subway.favorite.FavoriteRequest;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
+import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.domain.Member;
