+++ src/main/java/nextstep/subway/favorite/dto/FavoriteRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.favorite;
+package nextstep.subway.favorite.dto;
 
 public class FavoriteRequest {
     private long sourceStationId;
